requirements.txt
# Core Framework
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-chroma==0.1.4
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5

chromadb==0.5.23
openai==1.59.3
tiktoken==0.8.0
python-dotenv==1.0.1
pymupdf==1.25.1
sentence-transformers==3.3.1
rank_bm25==0.2.2
streamlit==1.41.1
ragas
